backend/requirements.txt
django
djangorestframework
Pillow
pandas
openpyxl
